ppo run

+++ realhf/impl/dataset/prompt_dataset.py
@@ -66,7 +66,7 @@ class PromptDataset(torch.utils.data.Dataset):
             trailing_shapes=dict(packed_prompts=()),
             dtypes=dict(packed_prompts=torch.long),
             ids=[self.ids[idx]],
-            seqlens=dict(packed_prompts=[torch.tensor([self.prompt_lengths[idx]])]),
+            seqlens=dict(packed_prompts=[torch.tensor([self.prompt_lengths[idx]], dtype=torch.int32)]),
             data=dict(packed_prompts=torch.tensor(self.prompts[idx], dtype=torch.long)),
         )
